refactor(operator): Pull out configuration of webhooks into k8s package.
